% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samc.R
\name{calc_b}
\alias{calc_b}
\title{Calculate B}
\usage{
calc_b(amc)
}
\arguments{
\item{amc}{Absorbing markov chain matrix. This should be output from the samc() function.}
}
\value{
B matrix
}
\description{
Calculate B
}
