% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samc.R
\name{calc_d}
\alias{calc_d}
\title{Calculate D}
\usage{
calc_d(amc)
}
\arguments{
\item{amc}{Absorbing markov chain matrix. This should be output from the samc() function.}
}
\value{
D matrix
}
\description{
Calculate D
}
