% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samc.R
\name{calc_disp}
\alias{calc_disp}
\title{Calculate dispersal}
\usage{
calc_disp(amc, pv)
}
\arguments{
\item{amc}{Absorbing markov chain matrix. This should be output from the samc() function.}

\item{pv}{Patch raster or vector}
}
\value{
disp
}
\description{
Caculate the product of the patch vector and D
}
