% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samc.R
\name{calc_trans_b}
\alias{calc_trans_b}
\title{Calculate transient B}
\usage{
calc_trans_b(amc, t)
}
\arguments{
\item{amc}{Absorbing markov chain matrix. This should be output from the samc() function.}
}
\value{
B matrix
}
\description{
Calculate transient B
}
