% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samc.R
\name{calc_trans_mort}
\alias{calc_trans_mort}
\title{Calculate transient mortatility}
\usage{
calc_trans_mort(amc, t, pv)
}
\arguments{
\item{amc}{Absorbing markov chain matrix. This should be output from the samc() function.}

\item{t}{Time}

\item{pv}{Patch raster or vector}
}
\value{
mort
}
\description{
Caculate the product of the patch vector and transient B
}
