% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samc.R
\name{calc_z}
\alias{calc_z}
\title{Calculate z}
\usage{
calc_z(amc)
}
\arguments{
\item{a}{Absorbing markov chain matrix. This should be output from the samc() function.}
}
\value{
z
}
\description{
Caculates z directly without the fundamental matrix, substantially reducing the memory requirements. 1,000,000 nodes should be feasible with 8GB of ram
}
