% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samc.R
\name{check_samc}
\alias{check_samc}
\title{Check if input is of class 'samc'}
\usage{
check_samc(amc)
}
\arguments{
\item{amc}{Absorbing markov chain matrix. This should be output from the samc() function.}
}
\value{
None
}
\description{
Check if input is of class 'samc'
}
