% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\docType{methods}
\name{dispersal}
\alias{dispersal}
\alias{dispersal,samc,missing,missing,numeric,numeric-method}
\alias{dispersal,samc,RasterLayer,missing,numeric,numeric-method}
\alias{dispersal,samc,matrix,missing,numeric,numeric-method}
\alias{dispersal,samc,missing,missing,missing,missing-method}
\alias{dispersal,samc,missing,numeric,missing,missing-method}
\alias{dispersal,samc,missing,missing,numeric,missing-method}
\alias{dispersal,samc,missing,numeric,numeric,missing-method}
\alias{dispersal,samc,RasterLayer,missing,missing,missing-method}
\alias{dispersal,samc,matrix,missing,missing,missing-method}
\alias{dispersal,samc,RasterLayer,missing,numeric,missing-method}
\alias{dispersal,samc,matrix,missing,numeric,missing-method}
\title{Calculate dispersal metrics}
\usage{
dispersal(samc, occ, origin, dest, time)

\S4method{dispersal}{samc,missing,missing,numeric,numeric}(samc, dest,
  time)

\S4method{dispersal}{samc,RasterLayer,missing,numeric,numeric}(samc, occ,
  dest, time)

\S4method{dispersal}{samc,matrix,missing,numeric,numeric}(samc, occ, dest,
  time)

\S4method{dispersal}{samc,missing,missing,missing,missing}(samc)

\S4method{dispersal}{samc,missing,numeric,missing,missing}(samc, origin)

\S4method{dispersal}{samc,missing,missing,numeric,missing}(samc, dest)

\S4method{dispersal}{samc,missing,numeric,numeric,missing}(samc, origin,
  dest)

\S4method{dispersal}{samc,RasterLayer,missing,missing,missing}(samc, occ)

\S4method{dispersal}{samc,matrix,missing,missing,missing}(samc, occ)

\S4method{dispersal}{samc,RasterLayer,missing,numeric,missing}(samc, occ,
  dest)

\S4method{dispersal}{samc,matrix,missing,numeric,missing}(samc, occ, dest)
}
\arguments{
\item{samc}{A \code{\link{samc-class}} object. This should be output from
the \code{\link{samc}} function.}

\item{occ}{A \code{\link[raster]{RasterLayer-class}} or \code{\link[base]{matrix}}.
The input type must match the input type used to create the
\code{\link{samc-class}} object, and must have the same properties as the
rest of the landscape data. See the \code{\link{check}} function for more
details.}

\item{origin}{A positive integer representing a cell in the landscape,
excluding \code{NA} cells. Corresponds to row \emph{i} of matrix \emph{P}
in the \code{\link{samc-class}} object.}

\item{dest}{A positive integer representing a cell in the landscape,
excluding \code{NA} cells. Corresponds to column \emph{j} of matrix \emph{P}
in the \code{\link{samc-class}} object.}

\item{time}{A positive integer representing time steps}
}
\value{
A matrix, a vector, or a numeric
}
\description{
Calculates the probability of individuals visiting locations
}
\details{
\eqn{\tilde{D}_{jt}=(\sum_{n=0}^{t-1}\tilde{Q}^n)\tilde{q}_j}
\itemize{
  \item \strong{dispersal(samc, dest, time)}

The result is a vector where each element corresponds to a cell in the
landscape, and can be mapped back to the landscape using the
\code{\link{map}} function. Element \emph{k} is the probability of ever
visiting a given destination, if starting at any other location, within
\emph{t} or fewer time steps.
}

\eqn{\psi^T\tilde{D}_{jt}}
\itemize{
  \item \strong{dispersal(samc, occ, dest, time)}

The result is a numeric that is the unconditional probability of visiting a
given destination within \emph{t} or fewer time steps.
}

\eqn{D=(F-I)diag(F)^{-1}}
\itemize{
  \item \strong{dispersal(samc)}

The result is a matrix where element (\emph{i},\emph{j}) is the probability
that location \emph{j} is visited when starting in location \emph{i}.

The returned matrix will always be dense and cannot be optimized. Must enable
override to use.

\item \strong{dispersal(samc, origin)}

This function has not been optimized yet, and will not run.

\item \strong{dispersal(samc, dest)}

The result is a vector where each element corresponds to a cell in the
landscape, and can be mapped back to the landscape using the
\code{\link{map}} function. Element \emph{i} is the probability that the
destination is visited when starting in location \emph{i}.

\item \strong{dispersal(samc, origin, dest)}

The result is a numeric value that is the probability that an individual
starting at the origin visits the destination.
}

\eqn{\psi^TD}
\itemize{
  \item \strong{dispersal(samc, occ)}

The result is a vector where each element corresponds to a cell in the
landscape, and can be mapped back to the landscape using the
\code{\link{map}} function. Element \emph{j} is the unconditional
probability distribution of ever visiting location \emph{j}, regardless of
the initial location.

\item \strong{dispersal(samc, occ, dest)}

The result is a numeric value that is the unconditional probability
distribution of ever visiting a given destination, regardless of the initial
location.
}
}
\section{Performance}{


Any relevant performance information about this function can be found in the
performance vignette: \href{../doc/performance.html}{\code{vignette("performance", package = "samc")}}
}

