% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_abs_data}
\alias{ex_abs_data}
\title{Example absorption data}
\format{An object of class \code{matrix} with 34 rows and 202 columns.}
\usage{
ex_abs_data
}
\description{
A matrix containing absorption probability values
}
\keyword{datasets}
