% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_mort_data}
\alias{ex_mort_data}
\title{Example mortality data}
\format{An object of class \code{matrix} with 120 rows and 240 columns.}
\usage{
ex_mort_data
}
\description{
A matrix containing mortality probability values
}
\keyword{datasets}
