% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_patch_data}
\alias{ex_patch_data}
\title{Example patch data}
\format{An object of class \code{matrix} with 120 rows and 240 columns.}
\usage{
ex_patch_data
}
\description{
A matrix containing starting location information
}
\keyword{datasets}
