% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_maze}
\alias{example_maze}
\title{Simple maze example}
\format{
A 20x20 matrix.
}
\source{
Generated for Maze Example vignettes
}
\usage{
example_maze
}
\description{
A simple maze layout.
}
\keyword{datasets}
