% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_split_corridor}
\alias{example_split_corridor}
\title{Simple split corridor example}
\format{
A list of three matrices. Matrices have 34 rows and 202 columns.
}
\source{
Fletcher et al (2019) <doi:10.1111/ele.13333>
}
\usage{
example_split_corridor
}
\description{
A fabricated dataset of a simple split corridor. Contains resistance,
absorption, and occupancy data.
}
\keyword{datasets}
