% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_toy_res}
\alias{example_toy_res}
\title{Toy resistance data}
\format{
A 10x10 matrix.
}
\source{
IALE 2021 workshop
}
\usage{
example_toy_res
}
\description{
Toy resistance data used in a workshop at IALE 2021
}
\keyword{datasets}
