% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samc.R
\name{fundamental}
\alias{fundamental}
\title{Calculate the fundamental matrix}
\usage{
fundamental(amc)
}
\arguments{
\item{amc}{Absorbing markov chain matrix. This should be output from the samc() function.}
}
\value{
Fundamental matrix.
}
\description{
Calculate the fundamental matrix
}
