% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location-class.R
\docType{class}
\name{location-class}
\alias{location-class}
\title{location class}
\description{
Union class for location inputs
}
\details{
The location class is a union class of the "numeric" and "character" classes.
Users generally do not need to worry about it except to know that any method
parameter with "location" as the type can have either an integer or a character
name provided as input.
}
