% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{samc-class-access}
\alias{samc-class-access}
\alias{$,samc-method}
\title{Access samc-class components}
\usage{
\S4method{$}{samc}(x, name)
}
\arguments{
\item{x}{samc-class object}

\item{name}{Component of the samc-class to access}
}
\value{
Component value
}
\description{
Allows users to access a subset of the samc-class components
}
\keyword{internal}
