% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{samc-class-modify}
\alias{samc-class-modify}
\alias{$<-,samc-method}
\title{Modify samc-class components}
\usage{
\S4method{$}{samc}(x, name) <- value
}
\arguments{
\item{x}{samc-class object}

\item{name}{Component of the samc-class to modify}

\item{value}{Value to assign to samc-class component}
}
\value{
Updated samc-class object
}
\description{
Allows users to modify a subset of the samc-class components
}
\keyword{internal}
