% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-classes.R
\docType{class}
\name{samc_data-class}
\alias{samc_data-class}
\title{data class}
\description{
Contains the data fields used in the samc-class
}
\section{Slots}{

\describe{
\item{\code{f}}{F matrix}

\item{\code{t_abs}}{Total absorption}

\item{\code{c_abs}}{Component absorption states}
}}

\keyword{internal}
