% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitation.R
\docType{methods}
\name{visitation}
\alias{visitation}
\alias{visitation,samc,missing,missing-method}
\alias{visitation,samc,numeric,missing-method}
\alias{visitation,samc,missing,numeric-method}
\alias{visitation,samc,numeric,numeric-method}
\title{Calculate visitation metrics}
\usage{
visitation(samc, origin, dest)

\S4method{visitation}{samc,missing,missing}(samc)

\S4method{visitation}{samc,numeric,missing}(samc, origin)

\S4method{visitation}{samc,missing,numeric}(samc, dest)

\S4method{visitation}{samc,numeric,numeric}(samc, origin, dest)
}
\arguments{
\item{samc}{A \code{\link{samc-class}} object. This should be output from
the \code{\link{samc}} function.}

\item{origin}{A positive integer representing a cell in the landscape,
excluding \code{NA} cells. Corresponds to row \emph{i} of matrix \emph{P}
in the \code{\link{samc-class}} object.}

\item{dest}{A positive integer representing a cell in the landscape,
excluding \code{NA} cells. Corresponds to column \emph{j} of matrix \emph{P}
in the \code{\link{samc-class}} object.}
}
\value{
A matrix, a vector, or a numeric
}
\description{
Calculates the number of times that individuals from each location visit each
location in the landscape before death.
}
\details{
\eqn{F = (I-Q)^{-1}}
\itemize{
  \item \strong{visitation(samc)}

The result is a matrix where element (\emph{i},\emph{j}) is the expected
number of times an individual that starts in \emph{i} uses \emph{j} before it
dies.

The returned matrix will always be dense and cannot be optimized. Must enable
override to use.

\item \strong{visitation(samc, origin)}

The result is a vector where each element corresponds to a cell in the
landscape, and can be mapped back to the landscape using the
\code{\link{map}} function. Element \emph{j} is the number of times that an
individual starting at the origin visits location \emph{j} before it dies.

\item \strong{visitation(samc, dest)}

The result is a vector where each element corresponds to a cell in the
landscape, and can be mapped back to the landscape using the
\code{\link{map}} function. Element \emph{i} is the number of times that an
individual starting at location \emph{i} visits the destination before it
dies.

\item \strong{visitation(samc, origin, dest)}

The result is a numeric value that is the expected number of times an
individual starting at the origin visits the destination before it dies.
}
}
\section{Performance}{


Any relevant performance information about this function can be found in the
performance vignette: \href{../doc/performance.html}{\code{vignette("performance", package = "samc")}}
}

