% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\docType{methods}
\name{check}
\alias{check}
\alias{check,RasterLayer,missing-method}
\alias{check,matrix,missing-method}
\alias{check,RasterLayer,RasterLayer-method}
\alias{check,matrix,matrix-method}
\alias{check,samc,RasterLayer-method}
\alias{check,samc,matrix-method}
\title{Check landscape data}
\usage{
check(a, b)

\S4method{check}{RasterLayer,missing}(a)

\S4method{check}{matrix,missing}(a)

\S4method{check}{RasterLayer,RasterLayer}(a, b)

\S4method{check}{matrix,matrix}(a, b)

\S4method{check}{samc,RasterLayer}(a, b)

\S4method{check}{samc,matrix}(a, b)
}
\arguments{
\item{a}{A \code{\link{samc-class}}, \code{\link{matrix}}, or \code{\link[raster]{RasterLayer-class}} object}

\item{b}{A \code{\link{matrix}} or \code{\link[raster]{RasterLayer-class}} object}
}
\value{
See \emph{Details} section.
}
\description{
Check that landscape inputs have valid values and matching properties.
}
\details{
This function is used to ensure that landscape inputs (resistance, absorption,
fidelity, and occupancy) have valid values and the same properties. This includes
checking the CRS (if using RasterLayer inputs), dimensions, and locations of
cells with NA data. It can be used to directly compare two matrices or two
RasterLayers, or it can be used to check a \code{\link{samc-class}} object
against a matrix or RasterLayer.

The function returns \code{TRUE} if the inputs have matching properties. Otherwise,
it will stop execution and print the error message generated by the
compareRaster() function from the raster package. This error will provide
some details about the difference between the two inputs.

Note that the package assumes the different landscape inputs will be the same
type, either matrices or RasterLayers. Mixing RasterLayer data and matrix
data is not supported.
}
