% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_abs_data}
\alias{ex_abs_data}
\title{Example absorption data}
\format{
A matrix with 34 rows and 202 columns.
}
\source{
Fletcher et al (2019) <doi:10.1111/ele.13333>
}
\usage{
ex_abs_data
}
\description{
A fabricated dataset containing landscape absorption probability values.
}
\keyword{datasets}
