% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_res_data}
\alias{ex_res_data}
\title{Example resistance data}
\format{An object of class \code{matrix} with 34 rows and 202 columns.}
\usage{
ex_res_data
}
\description{
A matrix containing movement resistance values
}
\keyword{datasets}
