% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{methods}
\name{map}
\alias{map}
\alias{map,samc,vector-method}
\title{Map vector data}
\usage{
map(samc, vec)

\S4method{map}{samc,vector}(samc, vec)
}
\arguments{
\item{samc}{Spatial absorbing markov chain object. This should be output from the samc() function.}

\item{vec}{Vector data to fill into the map.}
}
\value{
A RasterLayer object
}
\description{
Map vector data to a RasterLayer
}
\details{
This is a convenience function to ensure that vector data is properly mapped
back to the original landscape data. The reason this is needed is that the
package supports both matrices and RasterLayers, which differ in the order
that data is read and written (R matrices are column-major order, whereas the
raster package uses row-major order). Internally, the package uses only a
single order, regardless of the original data. This can cause issues with
mapping vector results if care is not taken, and this function is provided to
simplify the process. It also correctly maps results for landscape data that
has NA cells, which are another potential source of error if not careful.

The only requirement of the \code{vec} input is that the number of elements
in it matches the number of non-NA cells in the landscape data that was used
to create the samc object.
}
