% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samc.R
\name{samc}
\alias{samc}
\title{Create an absorbing Markov chain matrix}
\usage{
samc(resistance = NA, absorption = NA, latlon = NA, tr_fun = mean)
}
\arguments{
\item{resistance}{A RasterLayer object.}

\item{absorption}{A RasterLayer object.}
}
\value{
A spatial absorbing Markov chain matrix.
}
\description{
Create an absorbing Markov chain matrix
}
