% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\docType{methods}
\name{survival}
\alias{survival}
\alias{survival,samc,missing-method}
\alias{survival,samc,RasterLayer-method}
\alias{survival,samc,matrix-method}
\title{Calculate survival metrics}
\usage{
survival(samc, occ)

\S4method{survival}{samc,missing}(samc)

\S4method{survival}{samc,RasterLayer}(samc, occ)

\S4method{survival}{samc,matrix}(samc, occ)
}
\arguments{
\item{samc}{A \code{\link{samc-class}} object. This should be output from
the \code{\link{samc}} function.}

\item{occ}{A \code{\link[raster]{RasterLayer-class}} or \code{\link[base]{matrix}}.
The input type must match the input type used to create the
\code{\link{samc-class}} object, and must have the same properties as the
rest of the landscape data. See the \code{\link{check}} function for more
details.}
}
\value{
A \code{vector} or a \code{numeric}
}
\description{
Calculates the expected amount of time that individuals survive
in the landscape.
}
\details{
\eqn{z=(I-Q)^{-1}{\cdot}1=F{\cdot}1}
\itemize{
  \item \strong{survival(samc)}

The result is a vector where each element corresponds to a cell in the
landscape, and can be mapped back to the landscape using the
\code{\link{map}} function. The value of element \emph{i} is the expected
amount of time that individuals survive when starting at location \emph{j}.
}

\eqn{\psi^Tz}
\itemize{
  \item \strong{survival(samc, occ)}

The result is a numeric that represents the expected time that any individual
stays in the landscape before death, regardless of the initial location.
}
}
\section{Performance}{


Any relevant performance information about this function can be found in the
performance vignette: \href{../doc/performance.html}{\code{vignette("performance", package = "samc")}}
}

